+++ Fabric/SpatialHarvesters_1.17.x/src/main/java/dunkmania101/spatialharvesters/objects/tile_entities/DarkMobHarvesterTE.java
@@ -59,12 +59,12 @@ public class DarkMobHarvesterTE extends MobHarvesterTE {
     }
 
     @Override
-    public double getPrice(Block block) {
+    public long getPrice(Block block) {
         return CommonConfig.price_dark_mob;
     }
 
     @Override
-    public double getSpeed(Block block) {
+    public int getSpeed(Block block) {
         return CommonConfig.speed_dark_mob;
     }
